+++ packages/core/test/test-util/filter-factory.ts
@@ -66,6 +66,17 @@ export const testFilterLambdaExpressionOnLink = {
     "(to_MultiLink/any(a0:(a0/StringProperty eq 'test1')) and to_MultiLink/all(a0:(a0/StringProperty eq 'test2')))"
 };
 
+export const testFilterLambdaExpressionWithOr = {
+  filter: or(
+    TestEntityV4.STRING_PROPERTY.equals('str1'),
+    TestEntityV4.TO_MULTI_LINK.filter(
+      any(TestEntityMultiLinkV4.STRING_PROPERTY.equals('str2'))
+    )
+  ),
+  odataStr:
+    "(StringProperty eq 'str1' or (to_MultiLink/any(a0:(a0/StringProperty eq 'str2'))))"
+};
+
 export const testFilterLambdaExpressionFilterListOnLink = {
   filter: TestEntityV4.TO_MULTI_LINK.filter(
     any(